coderhub 参数错误bug优化

+++ src/router/upload/middleware.js
@@ -5,13 +5,15 @@ const multer = require('koa-multer')
 
 const {
   getInfo,
-  remove,
-  uploadAvatar,
+  updateAvatar,
+  createAvatar,
   updateAvatarUrl,
+  remove,
   uploadPictures
 } = require('./service')
 const { AVATAR_PATH, PICTURE_PATH } = require('../../util/file-path')
-const { APP_URL } = require('../../app/config')
+const { PARAMS_ERROR } = require('../../util/error-type')
+const { APP_URL, APP_PORT } = require('../../app/config')
 
 const saveAvatarMulter = multer({ dest: AVATAR_PATH })
 const savePicturesMulter = multer({ dest: PICTURE_PATH })
@@ -26,58 +28,73 @@ class UploadMiddleware {
   // 处理头像信息
   async handleAvatar(ctx, next) {
     const { id } = ctx.user
+
+    if(!ctx.req.file) return ctx.app.emit('error', new Error(PARAMS_ERROR), ctx)
     const { filename, mimetype, size } = ctx.req.file
+    
+    try {
+      // 获取原头像信息
+      const result = await getInfo("avatar", "user_id", id)
 
-    // 获取原头像信息（用于服务器删除图片）
-    const result = await getInfo("avatar", "user_id", id)
-
-    // 清除原头像
-    await remove("avatar", "user_id", id)
-
-    // 上传新头像
-    await uploadAvatar(id, filename, mimetype, size)
-
-    // 更新用户头像信息
-    await updateAvatarUrl(`${APP_URL}/user/${id}/avatar`, id)
-
-    // 服务器删除原头像
-    fs.unlink(path.join(AVATAR_PATH, `/${result[0].filename}`), error => {
-      if(error) {
-        console.log("用户头像删除失败", error);
+      if(result.length) {
+        // 更新头像
+        await updateAvatar(id, filename, mimetype, size)
+  
+        // 服务器删除原头像
+        fs.unlink(path.join(AVATAR_PATH, `/${result[0].filename}`), error => {
+          if(error) {
+            console.log("用户头像删除失败", error);
+          }
+        })
+      }else {
+        // 新建头像
+        await createAvatar(id, filename, mimetype, size)
       }
-    })
 
-    ctx.body = "上传头像成功~"
+      // 更新用户信息的头像地址
+      await updateAvatarUrl(`${APP_URL}:${APP_PORT}/user/${id}/avatar`, id)
+      ctx.body = "上传头像成功~"
+    } catch (error) {
+      ctx.body = error
+    }
   }
 
   // 处理配图信息
   async handlePictures(ctx, next) {
     const { id } = ctx.user
+
+    if(!ctx.req.files || !ctx.params) return ctx.app.emit('error', new Error(PARAMS_ERROR), ctx)
     const files = ctx.req.files
     const { momentId } = ctx.params
 
-    // 获取原配图信息
-    const result = await getInfo("picture", "moment_id", momentId)
+    try {
+      // 获取原配图信息
+      const result = await getInfo("picture", "moment_id", momentId)
 
-    // 清除原动态配图
-    await remove("picture", "moment_id", momentId)
+      if(result.length) {
+        // 清除所有原动态配图
+        await remove(momentId)
 
-    // 上传配图
-    for(let file of files) {
-      const { filename, mimetype, size } = file
-      await uploadPictures(id, filename, mimetype, size, momentId)
-    }
-
-    // 服务器删除原配图
-    for(let file of result) {
-      fs.unlink(path.join(PICTURE_PATH, `/${file.filename}`), error => {
-        if(error) {
-          console.log("配图删除失败", error);
+        // 服务器删除原配图
+        for(let file of result) {
+          fs.unlink(path.join(PICTURE_PATH, `/${file.filename}`), error => {
+            if(error) {
+              console.log("配图删除失败", error);
+            }
+          })
         }
-      })
+      }
+      
+      // 上传新配图信息
+      for(let file of files) {
+        const { filename, mimetype, size } = file
+        await uploadPictures(id, filename, mimetype, size, momentId)
+      }
+
+      ctx.body = "上传配图成功~"
+    } catch (error) {
+      ctx.body = error
     }
-    
-    ctx.body = "上传配图成功~"
   }
 }
